services/backend-core/requirements.txt
# FastAPI and Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# AI/ML Libraries
torch>=2.0.0
torchvision>=0.15.0
onnxruntime-gpu>=1.16.0
opencv-python>=4.9.0
pillow>=10.0.0
numpy>=1.26.0
scikit-learn>=1.3.0

# MLflow
mlflow>=2.10.0
boto3>=1.34.0

# Image Processing
albumentations>=1.3.0

# Vector Search
faiss-cpu>=1.9.0

# Authentication
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0

# Monitoring
prometheus-client>=0.19.0

# TAS (System 이상발생 분석 관리)
python-pptx>=0.6.23
pdfplumber>=0.11.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
